Refactors team analysis logger to team utils and applies on dev builds.
