+++ src/app/[locale]/layout.tsx
@@ -23,7 +23,8 @@ export const generateStaticParams: GenerateStaticParamsFunc<LocaleLayoutParams> 
   return locales.map((locale) => ({locale}));
 };
 
-const isLoadReactDevTools = 'true' === process.env?.REACT_APP_DEV_TOOLS;
+const isLoadReactDevTools = ! isProduction && 'true' === process.env?.REACT_APP_DEV_TOOLS;
+
 const font = Noto_Sans({
   weight: '400',
   subsets: ['latin'],
@@ -44,7 +45,7 @@ const RootLayout = ({children, params}: React.PropsWithChildren<LocaleLayoutProp
       {/* Google Analytics */}
       {isProduction() && (
         <>
-          <Script strategy="lazyOnload" src="https://www.googletagmanager.com/gtag/js?id=G-2LL7T4CCZP" />
+          <Script strategy="lazyOnload" src="https://www.googletagmanager.com/gtag/js?id=G-2LL7T4CCZP"/>
           <Script id="google-analytics">
             {`
               window.dataLayer = window.dataLayer || [];
@@ -58,9 +59,9 @@ const RootLayout = ({children, params}: React.PropsWithChildren<LocaleLayoutProp
           </Script>
         </>
       )}
-      {!isProduction() && isLoadReactDevTools && <Script src="http://localhost:8097"></Script>}
+      {isLoadReactDevTools && <Script src="http://localhost:8097"></Script>}
       <React.Suspense>
-        <SiteTracking />
+        <SiteTracking/>
       </React.Suspense>
       <body className={font.className}>
         <Providers>{children}</Providers>
